+++ components/cli/cmd/cellery/pull.go
@@ -21,10 +21,11 @@ package main
 import (
 	"fmt"
 
+	"github.com/cellery-io/sdk/components/cli/pkg/image"
+
 	"github.com/spf13/cobra"
 
 	"github.com/cellery-io/sdk/components/cli/pkg/commands"
-	"github.com/cellery-io/sdk/components/cli/pkg/util"
 )
 
 func newPullCommand() *cobra.Command {
@@ -38,7 +39,7 @@ func newPullCommand() *cobra.Command {
 			if err != nil {
 				return err
 			}
-			err = util.ValidateImageTagWithRegistry(args[0])
+			err = image.ValidateImageTagWithRegistry(args[0])
 			if err != nil {
 				return err
 			}
