+++ components/cli/cmd/cellery/push.go
@@ -21,10 +21,10 @@ package main
 import (
 	"fmt"
 
-	"github.com/spf13/cobra"
-
 	"github.com/cellery-io/sdk/components/cli/pkg/commands"
-	"github.com/cellery-io/sdk/components/cli/pkg/util"
+	"github.com/cellery-io/sdk/components/cli/pkg/image"
+
+	"github.com/spf13/cobra"
 )
 
 func newPushCommand() *cobra.Command {
@@ -38,7 +38,7 @@ func newPushCommand() *cobra.Command {
 			if err != nil {
 				return err
 			}
-			err = util.ValidateImageTagWithRegistry(args[0])
+			err = image.ValidateImageTagWithRegistry(args[0])
 			if err != nil {
 				return err
 			}
