+++ components/cli/cmd/cellery/test.go
@@ -22,11 +22,12 @@ import (
 	"fmt"
 	"regexp"
 
+	"github.com/cellery-io/sdk/components/cli/pkg/image"
+
 	"github.com/spf13/cobra"
 
 	"github.com/cellery-io/sdk/components/cli/pkg/commands"
 	"github.com/cellery-io/sdk/components/cli/pkg/constants"
-	"github.com/cellery-io/sdk/components/cli/pkg/util"
 )
 
 func newTestCommand() *cobra.Command {
@@ -45,7 +46,7 @@ func newTestCommand() *cobra.Command {
 			if err != nil {
 				return err
 			}
-			err = util.ValidateImageTagWithRegistry(args[0])
+			err = image.ValidateImageTagWithRegistry(args[0])
 			if err != nil {
 				return err
 			}
