+++ components/cli/pkg/commands/setup_create_gcp.go
@@ -61,7 +61,7 @@ func RunSetupCreateGcp(isCompleteSetup bool) {
 	} else {
 		createMinimalGcpRuntime()
 	}
-	util.WaitForRuntime(true, false)
+	runtime.WaitForRuntime(true, false)
 }
 
 func createGcp() error {
